feat(testing): split & upgrade community_base_test & testing utils
