+++ xieffect/test/communities/base/invitaion_test.py
@@ -12,119 +12,18 @@ from common.testing import SocketIOTestClient
 INVITATIONS_PER_REQUEST = 20
 
 
-def assert_create_community(socketio_client: SocketIOTestClient, community_data: dict):
-    ack = socketio_client.emit("new-community", community_data, callback=True)
-    events = socketio_client.get_received()
-    assert len(events) == 0
-    assert ack.get("code", None) == 200
-
-    result_data = ack.get("data", None)
-    assert result_data is not None
-
-    community_id = result_data.get("id", None)
-    assert isinstance(community_id, int)
-    assert dict_equal(result_data, community_data, *community_data.keys())
-    return community_id
-
-
-def get_communities_list(client: FlaskClient):
-    result = check_code(client.get("/home/")).get("communities", None)
-    assert isinstance(result, list)
-    return result
-
-
-@mark.order(1000)
-def test_meta_creation(client: FlaskClient, socketio_client: SocketIOTestClient):
-    community_ids = [d["id"] for d in get_communities_list(client)]
-
-    community_data = {"name": "12345", "description": "test"}
-    community_id = assert_create_community(socketio_client, community_data)
-    community_ids.append(community_id)
-
-    found = False
-    for data in get_communities_list(client):
-        assert data["id"] in community_ids
-        if data["id"] == community_id:
-            assert not found
-            assert dict_equal(data, community_data, "name", "description")
-            found = True
-    assert found
-
-
-@mark.order(1005)
-def test_community_list(client: FlaskClient, socketio_client: SocketIOTestClient):
-    def assert_order():
-        for i, data in enumerate(get_communities_list(client)):
-            assert data["id"] == community_ids[i]
-
-    socketio_client2 = SocketIOTestClient(client)
-    community_ids = [d["id"] for d in get_communities_list(client)]
-    assert_order()
-
-    # TODO check order with new community listing
-
-    # Creating
-    def assert_double_create(community_data: dict):
-        community_id = assert_create_community(socketio_client, community_data)
-
-        events = socketio_client2.get_received()
-        assert len(events) == 1
-        assert (event := events[0])["name"] == "new-community"
-        assert len(args := event["args"]) == 1
-
-        assert len(data := args[0]) == len(community_data) + 1
-        assert data.get("id", None) == community_id
-        assert dict_equal(data, community_data, *community_data.keys())
-
-        return community_id
-
-    community_datas: list[dict[str, str | int]] = [
-        {"name": "12345"}, {"name": "54321", "description": "hi"}, {"name": "test", "description": "i"}
-    ]
-
-    for community_data in community_datas:
-        community_data["id"] = assert_double_create(community_data)
-        community_ids.insert(0, community_data["id"])
-    # assert_order
-
-    # Reordering
-    reorder_data = {"source-id": community_datas[0]["id"], "target-index": 0}
-    ack = socketio_client2.emit("reorder-community", reorder_data, callback=True)
-    assert dict_equal(ack, {"code": 200, "message": "Success"}, ("code", "message"))
-    assert len(socketio_client2.get_received()) == 0
-
-    events = socketio_client.get_received()
-    assert len(events) == 1
-    assert (event := events[0])["name"] == "reorder-community"
-    assert len(args := event["args"]) == 1
-
-    assert len(data := args[0]) == 2
-    assert dict_equal(data, reorder_data, *reorder_data.keys())
-
-    community_ids.remove(reorder_data["source-id"])
-    community_ids.insert(reorder_data["target-index"], reorder_data["source-id"])
-    # assert_order
-
-    # Leaving
-    leave_data = {"community-id": community_datas[-1]["id"]}
-    ack = socketio_client.emit("leave-community", leave_data, callback=True)
-    assert dict_equal(ack, {"code": 200, "message": "Success"}, ("code", "message"))
-    assert len(socketio_client.get_received()) == 0
-
-    events = socketio_client2.get_received()
-    assert len(events) == 1
-    assert (event := events[0])["name"] == "leave-community"
-    assert len(args := event["args"]) == 1
-
-    assert len(data := args[0]) == 1
-    assert dict_equal(data, leave_data, *leave_data.keys())
-
-    community_ids.remove(leave_data["community-id"])
-    # assert_order
+@mark.order(1020)
+def test_invitations(client, list_tester, test_community):
+    invitation_data = {
+        "community-id": test_community,
+        "role": "base",
+        "limit": 2,
+        "days": 10,
+    }
+    room_data = {"community-id": test_community}
 
 
 @mark.skip
-@mark.order(1020)
 def test_invitations(client: FlaskClient, list_tester: Callable[[str, dict, int], Iterator[dict]]):
     community_data = {"name": "test", "description": "12345"}
